+++ README.md
@@ -14,6 +14,7 @@
   - [] 구매한 로또들의 번호 목록 
   - [] 당첨 번호 입력 안내 메세지 
   - [] 당첨 통계 
+    - [] 수익률은 소수점 아래 3자리까지만 나타낸다  
 - 로또 게임 
   - [x] 입력된 금액에 따라 로또를 구매한다 
     - [x] 천원 단위가 아니면 예외를 발생 -> 추후 게임 진행 시 재입력을 유도한다
@@ -27,4 +28,3 @@
   - [x] 등수 별로 개수를 기록한다
 - 로또 통계 
   - [x] 구매한 금액 대비 당첨 금액으로 수익률을 계산한다 
-  - [] 수익률은 소수점 아래 3자리까지만 나타낸다
