+++ src/main/kotlin/lotto/domain/InvalidLottoNumberCountException.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class InvalidLottoNumberCountException(numbers: Set<Int>) : RuntimeException(
     "로또 번호는 ${Lotto.NUMBER_COUNT}개가 입력되어야 합니다. 현재 입력 = $numbers",
