+++ src/main/kotlin/lotto/domain/InvalidMatchCountException.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.domain
 
 class InvalidMatchCountException(matchCount: Int) : RuntimeException("matchCount=$matchCount 에 해당하는 LottoRank가 없습니다.")
