+++ src/main/kotlin/lotto/domain/LottoResults.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoResults(ranks: List<LottoRank>) {
     val resultCountByRank: Map<LottoRank, Int> = ranks.groupingBy { it }.eachCount()
