+++ src/main/kotlin/lotto/domain/NotEnoughMoneyException.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.domain
 
 class NotEnoughMoneyException(money: Int) : RuntimeException("입력 금액은 1천원 단위여야 합니다. 현재 입력 = $money")
