+++ src/main/kotlin/lotto/domain/WinningNumbers.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class WinningNumbers(private val numbers: Set<Int>) {
     fun contains(number: Int): Boolean {
