+++ src/test/kotlin/lotto/domain/LottoProfitRateCalculatorTest.kt
@@ -1,7 +1,6 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.doubles.plusOrMinus
 import io.kotest.matchers.shouldBe
