+++ src/test/kotlin/lotto/domain/LottoRankTest.kt
@@ -1,8 +1,7 @@
-package lotto
+package lotto.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 
 class LottoRankTest : FreeSpec({
