+++ src/test/kotlin/lotto/domain/LottoResultCheckerTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
