+++ src/test/kotlin/lotto/domain/LottosTest.kt
@@ -1,8 +1,7 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
-import org.junit.jupiter.api.Assertions.*
 
 class LottosTest : FreeSpec({
     "로또 구매 가격을 계산한다" {
