+++ src/test/kotlin/lotto/domain/RandomLottoNumberGeneratorTest.kt
@@ -1,13 +1,8 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.core.spec.style.StringSpec
-import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.ints.shouldBeInRange
-import io.kotest.matchers.shouldBe
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
 
 class RandomLottoNumberGeneratorTest : FreeSpec({
     "6개의 숫자 목록을 생성한다" {
